third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/cycles-in-shadow-expected.txt

CONSOLE WARNING: line 12: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test that cycle detection traverses over both templates and shadow roots

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS shadowDiv.appendChild(outerDiv) threw exception HierarchyRequestError: Failed to execute 'appendChild' on 'Node': The new child element contains the parent..
PASS successfullyParsed is true

TEST COMPLETE
